# Implementation for alternative solution to gcc4.1 bug work-around