--- a/crypto/fipsmodule/ec/p256-nistz_test.cc
+++ b/crypto/fipsmodule/ec/p256-nistz_test.cc
@@ -40,33 +40,25 @@
     !defined(OPENSSL_SMALL) && !defined(BORINGSSL_SHARED_LIBRARY)
 
 TEST(P256_NistzTest, SelectW5) {
-  // Fill a table with some garbage input.
-  #if defined(__aarch64__)
-    // A bug on aarch not fixed in gcc 7 and 9 prevents the alignas usage 
-    // in cpp files where the argument to alignas is "seemingly" larger than the size of the array being aligned (faulty check).
-    // See references
-    // 1. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89357#c3
-    // 2. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=57271
-    // 3. https://github.com/facebook/folly/issues/1098
-    __attribute__((aligned(64))) P256_POINT table[16];
-  #else
-    alignas(64) P256_POINT table[16];
-  #endif
+  // Fill a table with some garbage input.  
+  char buffer_table[64 + (sizeof(P256_POINT) * 16)];
+  P256_POINT *aligned_table = (P256_POINT *) align_pointer(buffer_table, 64);
+
   for (size_t i = 0; i < 16; i++) {
-    OPENSSL_memset(table[i].X, 3 * i, sizeof(table[i].X));
-    OPENSSL_memset(table[i].Y, 3 * i + 1, sizeof(table[i].Y));
-    OPENSSL_memset(table[i].Z, 3 * i + 2, sizeof(table[i].Z));
+    OPENSSL_memset(aligned_table[i].X, 3 * i, sizeof(aligned_table[i].X));
+    OPENSSL_memset(aligned_table[i].Y, 3 * i + 1, sizeof(aligned_table[i].Y));
+    OPENSSL_memset(aligned_table[i].Z, 3 * i + 2, sizeof(aligned_table[i].Z));
   }
 
   for (int i = 0; i <= 16; i++) {
     P256_POINT val;
-    ecp_nistz256_select_w5(&val, table, i);
+    ecp_nistz256_select_w5(&val, aligned_table, i);
 
     P256_POINT expected;
     if (i == 0) {
       OPENSSL_memset(&expected, 0, sizeof(expected));
     } else {
-      expected = table[i-1];
+      expected = aligned_table[i-1];
     }
 
     EXPECT_EQ(Bytes(reinterpret_cast<const char *>(&expected), sizeof(expected)),
@@ -76,36 +68,28 @@ TEST(P256_NistzTest, SelectW5) {
   // This is a constant-time function, so it is only necessary to instrument one
   // index for ABI checking.
   P256_POINT val;
-  CHECK_ABI(ecp_nistz256_select_w5, &val, table, 7);
+  CHECK_ABI(ecp_nistz256_select_w5, &val, aligned_table, 7);
 }
 
 TEST(P256_NistzTest, SelectW7) {
   // Fill a table with some garbage input.
-  #if defined(__aarch64__)
-    // A bug on aarch not fixed in gcc 7 and 9 prevents the alignas usage 
-    // in cpp files where the argument to alignas is "seemingly" larger than the size of the array being aligned (faulty check).
-    // See references
-    // 1. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89357#c3
-    // 2. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=57271
-    // 3. https://github.com/facebook/folly/issues/1098
-    __attribute__((aligned(64))) P256_POINT_AFFINE table[64];
-  #else
-    alignas(64) P256_POINT_AFFINE table[64];
-  #endif
+  char buffer_table[64 + (sizeof(P256_POINT_AFFINE) * 64)];
+  P256_POINT_AFFINE *aligned_table = (P256_POINT_AFFINE *) align_pointer(buffer_table, 64);
+
   for (size_t i = 0; i < 64; i++) {
-    OPENSSL_memset(table[i].X, 2 * i, sizeof(table[i].X));
-    OPENSSL_memset(table[i].Y, 2 * i + 1, sizeof(table[i].Y));
+    OPENSSL_memset(aligned_table[i].X, 2 * i, sizeof(aligned_table[i].X));
+    OPENSSL_memset(aligned_table[i].Y, 2 * i + 1, sizeof(aligned_table[i].Y));
   }
 
   for (int i = 0; i <= 64; i++) {
     P256_POINT_AFFINE val;
-    ecp_nistz256_select_w7(&val, table, i);
+    ecp_nistz256_select_w7(&val, aligned_table, i);
 
     P256_POINT_AFFINE expected;
     if (i == 0) {
       OPENSSL_memset(&expected, 0, sizeof(expected));
     } else {
-      expected = table[i-1];
+      expected = aligned_table[i-1];
     }
 
     EXPECT_EQ(Bytes(reinterpret_cast<const char *>(&expected), sizeof(expected)),
@@ -115,7 +99,7 @@ TEST(P256_NistzTest, SelectW7) {
   // This is a constant-time function, so it is only necessary to instrument one
   // index for ABI checking.
   P256_POINT_AFFINE val;
-  CHECK_ABI(ecp_nistz256_select_w7, &val, table, 42);
+  CHECK_ABI(ecp_nistz256_select_w7, &val, aligned_table, 42);
 }
 
 TEST(P256_NistzTest, BEEU) {