### Description of changes: 

Solution that was discarded: https://github.com/awslabs/aws-lc/pull/312

Ancient gcc compilers (up to (not including) version 4.6) contain a bug that cause wrong alignment for stack allocated memory even though it was tagged with an appropriate specifier `alignas`/`align` attribute. The bug is explained in more depth here: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=16660

To work around this bug, we use a temporary byte buffer. Say we want a pointer `aligned_ptr` of type `type_aligned_ptr`. We then allocate a buffer of bytes, find the correct alignment and cast the remaining bytes to the desired type.

```
stack_align_type bufffer[align_boundary + size of object(s)]
// align pointer |buffer| to `align_boundary` boundary.
// use align_pointer() for this.
// Aligned pointer is `aligned_buffer_ptr`.
type_aligned_ptr *aligned_ptr = (type_aligned_ptr *) aligned_buffer_ptr
```

`stack_align_tyoe` is a type of size 8 bits such that arithmetic in pointer-space moves 1 byte per basic increment. So for this, we choose `uint8_t`. `type_aligned_ptr` is the type the object the aligned pointer points to. The remaining memory, after aligning using `align_pointer`, should be sufficient to hold the object(s) since we allocated an extra `align_boundary` bytes in the buffer array (in theory we could do with `align_boundary` - 1).

Strongly recommend using modern compilers.

### Testing:
Try out this alignment for different `align_boundary` and object sizes. All test cases are generated using a Python script and split in two because the virtual functions produced by a test fixture becomes too big.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.