@torben-hansen 2021-12-12
Performance sniff test. On the face, looks (marginally?) more stable than #312.

**With patch**
```
% for i in `seq 1 4`; do ./tool/bssl speed -filter "ECDH P-256"; done
Did 17000 ECDH P-256 operations in 1034097us (16439.5 ops/sec)
Did 17000 ECDH P-256 operations in 1021496us (16642.3 ops/sec)
Did 17000 ECDH P-256 operations in 1021955us (16634.8 ops/sec)
Did 17000 ECDH P-256 operations in 1021462us (16642.8 ops/sec)

% for i in `seq 1 4`; do ./tool/bssl speed -filter "ECDSA P-256"; done
Did 51000 ECDSA P-256 signing operations in 1005729us (50709.5 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1047675us (18135.4 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1009842us (50502.9 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1043844us (18202.0 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1005964us (50697.6 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1049384us (18105.9 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1005687us (50711.6 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1046303us (18159.2 ops/sec)
```

**Without patch**
```
% for i in `seq 1 4`; do ./tool/bssl speed -filter "ECDH P-256"; done
Did 17000 ECDH P-256 operations in 1023951us (16602.4 ops/sec)
Did 17000 ECDH P-256 operations in 1024468us (16594.0 ops/sec)
Did 17000 ECDH P-256 operations in 1042906us (16300.6 ops/sec)
Did 17000 ECDH P-256 operations in 1024713us (16590.0 ops/sec)

% for i in `seq 1 4`; do ./tool/bssl speed -filter "ECDSA P-256"; done
Did 51000 ECDSA P-256 signing operations in 1008289us (50580.7 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1046213us (18160.7 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1007030us (50644.0 ops/sec)
Did 18000 ECDSA P-256 verify operations in 1015836us (17719.4 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1009448us (50522.7 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1048538us (18120.5 ops/sec)
Did 51000 ECDSA P-256 signing operations in 1008348us (50577.8 ops/sec)
Did 19000 ECDSA P-256 verify operations in 1047263us (18142.5 ops/sec)
```